convert to int

+++ RectangleGraph.py
@@ -104,7 +104,7 @@ def arrayToGraph(array):
                 for x in array[i][j]:
                     G.add_edge(i, j, gid=x, energy=Bus_energy, time=Bus_cycle)
             else:
-                G.add_edge(i, j, gid=float(array[i][j]), energy=Bus_energy, time=Bus_cycle)
+                G.add_edge(i, j, gid=int(array[i][j], energy=Bus_energy, time=Bus_cycle)
     return G
